[FEATURE] Take release comment from Payload HEAD commit
This change removes the default comment "Automatic release built from Github. See %s for change log" and introduces a new behavior:

* If the `comment` setting is specified that value is used as comment
* If no `comment` setting is specified then Payload HEAD's commit message body is used as comment.

This also removes the support for a `%s` marker in the comment which inserted the change log link - to add such links, simply place them as part of the comment.

Close: #2