[FEATURE] Allow `extensionKey` to be passed as URL parameter
If `$_GET['extensionKey']` is provided it gets used as extension key. If not provided, the same-named settings value is checked. If that's also not provided, the original behavior of using the repository's name is used as fallback.

Also implemented strict validation of this extension key at the point where the directory name is resolved in order to avoid any potential security issues such as some clever faker trying to pass actual filesystem paths in the argument.